cache plugin with shims
