+++ pom.xml
@@ -147,7 +147,7 @@
     <properties>
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
-        <spark.version>3.1.0-SNAPSHOT</spark.version>
+        <spark.version>3.0.0</spark.version>
         <cuda.version>cuda10-1</cuda.version>
         <cudf.version>0.15-SNAPSHOT</cudf.version>
         <scala.binary.version>2.12</scala.binary.version>
